requirements.txt
django
gunicorn
Pillow
django-recaptcha>=3.0.0
